Refactor: Move default_models.py to framework
Moved `plugin/contrib/default_models.py` to `plugin/framework/default_models.py` because it is internal project code, not third-party contrib code. Updated imports in `plugin/framework/config.py` to point to the new location.
